+++ LCD.h
@@ -8,7 +8,7 @@
 extern uint32_t ILI9341_SPICounter;
 
 extern void LCD_startLCD();
-extern void ILI9341_drawColors_indexed(uint32_t x, uint32_t y, int16_t *rgbIndex, uint16_t totalPixels);
+extern void ILI9341_drawColors_indexed(uint32_t x, uint32_t y, int32_t *rgbIndex, uint16_t totalPixels);
 extern void ILI9341_fillScreen(uint32_t rgb);
 extern void LCD_drawPixel(uint16_t x, uint16_t y, uint32_t rgb);
 extern void LCD_update();
